requirements.txt
flask
google-cloud-bigquery
google-cloud-aiplatform
boto3
azure-ai-textanalytics
pytest
